requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
alembic==1.13.2
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asteroid-filterbanks==0.4.0
attrs==24.2.0
audioread==3.0.1
av==12.3.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
ctranslate2==4.3.1
cycler==0.12.1
decorator==5.1.1
discord==2.3.2
discord.py==2.4.0
docopt==0.6.2
einops==0.8.0
evdev==1.7.1
fakeyou==1.2.6
faster-whisper @ https://github.com/SYSTRAN/faster-whisper/archive/refs/heads/master.tar.gz#sha256=f51a51e06c6b170e9f471d58f200a98fd45eed91e11d0b2d1713e8417c31b02c
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
keyboard==0.13.5
kiwisolver==1.4.5
lazy_loader==0.4
Levenshtein==0.25.1
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.6
linkify-it-py==2.0.3
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdit-py-plugins==0.4.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
npyscreen==4.10.5
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
ollama==0.3.1
omegaconf==2.3.0
onnxruntime==1.18.1
optuna==3.6.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
primePy==1.3
protobuf==5.27.3
pyannote.audio==3.3.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
pygame==2.6.0
Pygments==2.18.0
pynput==1.7.7
pyparsing==3.1.2
PyTermGUI==7.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-xlib==0.33
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.1
scipy==1.14.0
semver==3.0.2
sentencepiece==0.2.0
setuptools==72.1.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.4.0
speechbrain==1.0.0
SQLAlchemy==2.0.32
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.0
torchmetrics==1.4.1
torchvision==0.17.0
tqdm==4.66.5
triton==2.2.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
Wave==0.0.2
wcwidth==0.2.13
yarl==1.9.4
